test(EuiDataGrid): move rendering unit tests back to `data_grid.test.tsx` and convert them to RTL
